+++ packages/kit/test/apps/basics/src/routes/data-sveltekit/preload-data/+page.svelte
@@ -1,0 +1,8 @@
+<a id="one" href="/data-sveltekit/preload-data/target" data-sveltekit-preload-data>one</a>
+
+<div data-sveltekit-preload-data>
+	<a id="two" href="/data-sveltekit/preload-data/target">two</a>
+	<a id="three" href="/data-sveltekit/preload-data/target" data-sveltekit-preload-data="off"
+		>three</a
+	>
+</div>