+++ packages/create-svelte/CHANGELOG.md
@@ -1,5 +1,19 @@
 # create-svelte
 
+## 2.0.0-next.195
+
+### Patch Changes
+
+- [breaking] Replace `data-sveltekit-prefetch` with `-preload-code` and `-preload-data` ([#7776](https://github.com/sveltejs/kit/pull/7776))
+
+- Upgrade to Playwright 1.28.1 ([#7696](https://github.com/sveltejs/kit/pull/7696))
+
+## 2.0.0-next.194
+
+### Patch Changes
+
+- fix playwright glob filter ([#7826](https://github.com/sveltejs/kit/pull/7826))
+
 ## 2.0.0-next.193
 
 ### Patch Changes